requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.0
pyautogui==0.9.54
pillow==10.0.1 